pyproject.toml
[tool.poetry]
name = "forex-api"
version = "0.1.0"
description = ""
authors = ["Kamaal <kamaal.f1@gmail.com>"]
readme = "README.md"
packages = [{include = "forex_api"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.91.0"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
google-cloud-firestore = "^2.9.1"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
